refactor: consolidate controller implementation
